add to cart complete

+++ database/migrations/2020_07_15_102626_create_carts_table.php
@@ -19,7 +19,7 @@ class CreateCartsTable extends Migration
             $table->unsignedBigInteger('order_id')->nullable();
             $table->unsignedBigInteger('user_id')->nullable();
             $table->string('ip')->nullable();
-            $table->float('price')->nullable();
+            $table->unsignedBigInteger('price')->nullable();
             $table->enum('status',['new','progress','delivered','cancel'])->default('new');
             $table->integer('quantity')->nullable();
             $table->unsignedBigInteger('amount')->nullable();
