+++ database/migrations/2020_07_23_143757_create_wishlists_table.php
@@ -18,9 +18,9 @@ class CreateWishlistsTable extends Migration
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('cart_id')->nullable();
             $table->unsignedBigInteger('user_id')->nullable();
-            $table->float('price');
+            $table->unsignedBigInteger('price');
             $table->integer('quantity');
-            $table->float('amount');
+            $table->unsignedBigInteger('amount');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('CASCADE');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('SET NULL');
             $table->foreign('cart_id')->references('id')->on('carts')->onDelete('SET NULL');
